move run_summary from tf_kernel to pipelines
