+++ tests/parkfield/test_process_parkfield_run_rr.py
@@ -5,11 +5,11 @@ from aurora.config.config_creator import ConfigCreator
 from aurora.config.metadata import Processing
 
 from aurora.pipelines.process_mth5 import process_mth5
+from aurora.pipelines.run_summary import RunSummary
 from aurora.test_utils.parkfield.path_helpers import AURORA_RESULTS_PATH
 from aurora.test_utils.parkfield.path_helpers import CONFIG_PATH
 from aurora.test_utils.parkfield.path_helpers import DATA_PATH
 from aurora.test_utils.parkfield.path_helpers import EMTF_RESULTS_PATH
-from aurora.tf_kernel.run_summary import RunSummary
 from aurora.transfer_function.kernel_dataset import KernelDataset
 from aurora.transfer_function.plot.comparison_plots import compare_two_z_files
