+++ tests/synthetic/test_stft_methods_agree.py
@@ -4,13 +4,13 @@ method returns the same array as scipy.signal.spectrogram
 """
 import numpy as np
 
+from aurora.pipelines.run_summary import RunSummary
 from aurora.pipelines.time_series_helpers import get_run_run_ts_from_mth5
 from aurora.pipelines.time_series_helpers import prototype_decimate
 from aurora.pipelines.time_series_helpers import run_ts_to_stft
 from aurora.pipelines.time_series_helpers import run_ts_to_stft_scipy
 from aurora.test_utils.synthetic.make_mth5_from_asc import create_test1_h5
 from aurora.test_utils.synthetic.make_processing_configs import create_test_run_config
-from aurora.tf_kernel.run_summary import RunSummary
 from aurora.transfer_function.kernel_dataset import KernelDataset
 from mth5.mth5 import MTH5
