+++ tests/synthetic/test_synthetic_driver.py
@@ -1,3 +1,4 @@
+from aurora.pipelines.run_summary import RunSummary
 from aurora.test_utils.synthetic.make_mth5_from_asc import create_test1_h5
 from aurora.test_utils.synthetic.make_mth5_from_asc import create_test2_h5
 from aurora.test_utils.synthetic.make_mth5_from_asc import create_test12rr_h5
@@ -5,7 +6,6 @@ from aurora.test_utils.synthetic.make_processing_configs import create_test_run_
 from aurora.test_utils.synthetic.paths import AURORA_RESULTS_PATH
 from aurora.test_utils.synthetic.paths import CONFIG_PATH
 from aurora.test_utils.synthetic.processing_helpers import process_sythetic_data
-from aurora.tf_kernel.run_summary import RunSummary
 from aurora.transfer_function.kernel_dataset import KernelDataset
 
 # def process_synthetic_1_underdetermined():
